chore(revert): remove rrweb
