+++ packages/session-recorder/src/recorder/index.ts
@@ -15,15 +15,9 @@
  * limitations under the License.
  *
  */
-
-export type RecorderType = 'rrweb' | 'splunk'
-
-export type SessionState = {
-	expiresAt: number
-	id: string
-	rt?: RecorderType
-	startTime: number
-	state: 'active' | 'native' | 'expired-inactivity' | 'expired-duration'
-}
-
-export type PersistedSessionState = Pick<SessionState, 'expiresAt' | 'rt' | 'id' | 'startTime'>
+export * from './recorder-base'
+export * from './rrweb-recorder'
+export * from './splunk-recorder'
+export * from './config'
+export * from './types'
+export * from './recorder'
