+++ packages/session-recorder/src/recorder/recorder-base.ts
@@ -15,11 +15,30 @@
  * limitations under the License.
  *
  */
-import { mergeRecorderConfig } from './migrations/config-merge'
-import { migrateRRWebConfigToSplunkConfig } from './migrations/config-migration'
-import { RecorderPublicConfig } from './recorder'
+export interface RecorderEmitContext {
+	data: Record<string, unknown>
+	startTime: number
+	type: 'splunk' | 'rrweb'
+}
 
-export const getSplunkRecorderConfig = (config: RecorderPublicConfig): RecorderPublicConfig => {
-	const migratedRRWebConfig = migrateRRWebConfigToSplunkConfig(config)
-	return mergeRecorderConfig(config, migratedRRWebConfig)
+export interface RecorderConfig {
+	onEmit: (context: RecorderEmitContext) => void
+}
+
+export abstract class RecorderBase {
+	protected onEmit: RecorderConfig['onEmit']
+
+	protected constructor(config: RecorderConfig) {
+		this.onEmit = config.onEmit
+	}
+
+	abstract onSessionChanged(): void
+
+	abstract pause(): void
+
+	abstract resume(): void
+
+	abstract start(): void
+
+	abstract stop(): void
 }
