+++ packages/session-recorder/src/recorder/types.ts
@@ -15,15 +15,4 @@
  * limitations under the License.
  *
  */
-
 export type RecorderType = 'rrweb' | 'splunk'
-
-export type SessionState = {
-	expiresAt: number
-	id: string
-	rt?: RecorderType
-	startTime: number
-	state: 'active' | 'native' | 'expired-inactivity' | 'expired-duration'
-}
-
-export type PersistedSessionState = Pick<SessionState, 'expiresAt' | 'rt' | 'id' | 'startTime'>
